# backend/app/routers/prices.py
from datetime import UTC, datetime
from typing import Annotated

from fastapi import APIRouter, HTTPException, Query, status
from sqlalchemy import or_
from sqlmodel import Session, select

from app.db.models import ListItem, ListMember, Purchase
from app.dependencies import CurrentSession, CurrentUser, MemberDep
from app.schemas.prices import PriceCreate, PriceEntry, PriceHistoryResponse
from app.services import trips
from app.services.store_registry import ensure_stores

router = APIRouter(prefix="/lists/{list_id}/items/{item_id}/prices", tags=["prices"])


def _get_item_or_404(session: Session, item_id: str, list_id: str) -> ListItem:
    item = session.exec(
        select(ListItem).where(ListItem.id == item_id, ListItem.list_id == list_id)
    ).first()
    if not item:
        raise HTTPException(status_code=404, detail="Item not found")
    return item


def _write_price(item: ListItem, price_in: PriceCreate, session: Session) -> PriceEntry:
    item.price = price_in.amount
    item.price_per = price_in.price_per
    item.price_store = price_in.store
    session.add(item)
    if price_in.store:
        ensure_stores(session, item.list_id, [price_in.store])
    session.commit()
    session.refresh(item)
    return PriceEntry(amount=item.price, price_per=item.price_per, store=item.price_store)


@router.post("", response_model=PriceEntry, status_code=status.HTTP_201_CREATED)
def create_price(
    list_id: str,
    item_id: str,
    price_in: PriceCreate,
    session: CurrentSession,
    current_user: CurrentUser,
    _: MemberDep,
):
    item = _get_item_or_404(session, item_id, list_id)
    if item.price is not None:
        raise HTTPException(
            status_code=409, detail="Item already has a price; use PATCH to update it"
        )
    return _write_price(item, price_in, session)


@router.patch("", response_model=PriceEntry)
def update_price(
    list_id: str,
    item_id: str,
    price_in: PriceCreate,
    session: CurrentSession,
    current_user: CurrentUser,
    _: MemberDep,
):
    item = _get_item_or_404(session, item_id, list_id)
    if item.price is None:
        raise HTTPException(status_code=404, detail="Item has no price yet; use POST to set it")
    return _write_price(item, price_in, session)


@router.delete("", status_code=status.HTTP_204_NO_CONTENT)
def delete_price(
    list_id: str,
    item_id: str,
    session: CurrentSession,
    current_user: CurrentUser,
    _: MemberDep,
):
    item = _get_item_or_404(session, item_id, list_id)
    if item.price is None:
        raise HTTPException(status_code=404, detail="Item has no price to delete")
    if item.purchased_at is not None:
        # A purchased item whose trip row is missing counts as closed —
        # refusing a delete is recoverable, erasing settled spend is not.
        now = datetime.now(UTC).replace(tzinfo=None)
        trip = session.get(Purchase, item.purchase_id) if item.purchase_id else None
        if trip is None or not trips.is_open(trip, now):
            raise HTTPException(
                status_code=status.HTTP_422_UNPROCESSABLE_CONTENT,
                detail="Cannot delete the price of an item from a closed trip",
            )
    item.price = None
    item.price_per = None
    item.price_store = None
    session.add(item)
    session.commit()


@router.get("", response_model=PriceHistoryResponse)
def get_price_history(
    list_id: str,
    item_id: str,
    session: CurrentSession,
    current_user: CurrentUser,
    _: MemberDep,
    scope: Annotated[str, Query(pattern="^(this_list|my_lists|all)$")] = "this_list",
):
    item = _get_item_or_404(session, item_id, list_id)
    rows = _query_by_scope(session, item, scope, current_user.id)
    entries = [
        PriceEntry(
            amount=i.price,
            price_per=i.price_per,
            # A sin-precio line has no price_store; its trip carries the store.
            store=i.price_store if i.price_store is not None else purchase_store,
            purchased_at=i.purchased_at.isoformat() if i.purchased_at else None,
            quantity=i.quantity,
            is_sin_precio=i.purchase_id is not None and i.price is None,
        )
        for i, purchase_store in rows
    ]
    # Dated rows first, oldest to newest; dateless manual logs trail behind.
    # purchased_at is ISO 8601, so string order is chronological order.
    entries.sort(key=lambda e: (e.purchased_at is None, e.purchased_at or ""))
    return PriceHistoryResponse(entries=entries)


def _query_by_scope(session, item: ListItem, scope: str, user_id: str) -> list:
    base = _base_conditions(item)
    # LEFT OUTER JOIN so priced manual logs (no trip) still come through, while
    # purchase-linked rows can read their trip's store for attribution.
    stmt = select(ListItem, Purchase.store).outerjoin(Purchase, ListItem.purchase_id == Purchase.id)

    if scope == "this_list":
        return session.exec(stmt.where(ListItem.list_id == item.list_id, *base)).all()

    if scope == "my_lists":
        my_list_ids = session.exec(
            select(ListMember.list_id).where(ListMember.user_id == user_id)
        ).all()
        return session.exec(stmt.where(ListItem.list_id.in_(my_list_ids), *base)).all()

    # scope == "all"
    return session.exec(stmt.where(*base)).all()


def _base_conditions(item: ListItem):
    # Surface purchase-linked OR priced rows: keeps manually-logged prices,
    # surfaces sin-precio purchase rows as gaps, and still excludes
    # pending-never-bought rows (both price and purchase_id null).
    surfaced = or_(ListItem.price.isnot(None), ListItem.purchase_id.isnot(None))
    if item.ean:
        return (ListItem.ean == item.ean, surfaced)
    return (ListItem.name == item.name, ListItem.brand == item.brand, surfaced)